+++ packages/react-ui/components/Hint/Hint.styles.ts
@@ -7,12 +7,20 @@ export const styles = memoizeStyle({
       box-sizing: border-box;
       color: ${t.hintColor};
       font-size: ${t.hintFontSize};
-      line-height: ${t.hintLineHeight}
+      line-height: ${t.hintLineHeight};
       max-width: ${t.hintMaxWidth};
       overflow-wrap: break-word;
       padding: ${t.hintPaddingY} ${t.hintPaddingX};
       word-break: break-word;
       word-wrap: break-word;
+    `;
+  },
+
+  mobileContent(t: Theme) {
+    return css`
+      color: ${t.mobileHintColor};
+      max-width: 100%;
+      text-align: left;
     `;
   },