+++ packages/react-ui/lib/locale/useLocaleForControl.ts
@@ -1,0 +1,14 @@
+import { useContext, useMemo } from 'react';
+
+import { LocaleControls } from './types';
+import { LocaleHelper } from './LocaleHelper';
+import { LocaleContext } from './LocaleContext';
+
+export const useLocaleForControl = <TTranslationByLangCode extends object>(
+  controlName: keyof LocaleControls,
+  localeHelper: LocaleHelper<TTranslationByLangCode>,
+): TTranslationByLangCode => {
+  const { locale, langCode } = useContext(LocaleContext);
+  const localeFromContext = locale?.[controlName];
+  return useMemo(() => Object.assign({}, localeHelper.get(langCode), localeFromContext), [langCode, localeFromContext]);
+};